+++ CHANGELOG.md
@@ -1,3 +1,31 @@
+# Changelog
+
+## <small>1.0.16 (2025-03-25)</small>
+
+* ci(release): 优化 NPM 发布流程 ([54e5959](https://github.com/h7ml/ai-markdown-translator/commit/54e5959))
+* build(release): 添加 release-it 配置并更新相关文件 ([ae6fd92](https://github.com/h7ml/ai-markdown-translator/commit/ae6fd92))
+* Fix: Path error for log ([9887d56](https://github.com/h7ml/ai-markdown-translator/commit/9887d56))
+* Fix: Print help only for main yargs process ([bc0e174](https://github.com/h7ml/ai-markdown-translator/commit/bc0e174))
+* Fix: Remove default open ai url while input args step ([047132e](https://github.com/h7ml/ai-markdown-translator/commit/047132e))
+* Fix: Roll back system role for compatibility ([c513857](https://github.com/h7ml/ai-markdown-translator/commit/c513857))
+* Update: CI config and reflect PR comment ([ba56eca](https://github.com/h7ml/ai-markdown-translator/commit/ba56eca))
+* Update: Cli description for locale flag ([6425aa0](https://github.com/h7ml/ai-markdown-translator/commit/6425aa0))
+* Update: Disable api type arg for now ([7b74154](https://github.com/h7ml/ai-markdown-translator/commit/7b74154))
+* Refactor: Clear name for two types of path ([38e781a](https://github.com/h7ml/ai-markdown-translator/commit/38e781a))
+* Refactor: Prepare before implement response API ([35a16c5](https://github.com/h7ml/ai-markdown-translator/commit/35a16c5))
+* Comment: For i18n.ts ([c7f87da](https://github.com/h7ml/ai-markdown-translator/commit/c7f87da))
+* Feat: Apply locale for cli help ([f5a75e8](https://github.com/h7ml/ai-markdown-translator/commit/f5a75e8))
+* Add: Example for .env configs ([0873def](https://github.com/h7ml/ai-markdown-translator/commit/0873def))
+* Hotfix: Wrong condition for use OpenAI module ([04d0672](https://github.com/h7ml/ai-markdown-translator/commit/04d0672))
+
+## [1.0.16](https://github.com/h7ml/ai-markdown-translator/compare/v1.0.15...v1.0.16) (2025-03-25)
+
+
+
+## [1.0.14](https://github.com/h7ml/ai-markdown-translator/compare/v1.0.15...v1.0.14) (2025-03-25)
+
+
+
 ## v1.0.13
 
 [compare changes](https://github.com/h7ml/ai-markdown-translator/compare/v1.0.13...v1.0.13)